requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lxml==5.4.0
Markdown==3.8
packaging==25.0
pluggy==1.5.0
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-docx==1.1.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
